Add files via upload
Datafile contains the edges of a directed graph G=(V,A) and we will use the data to create a shortest-path problem from the first to the last node (sorted by their index). The format of the data file is: i,j,c, where i and j are the endnodes of the directed i→j arc, while c is the cost of the arc.